backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
flask-bcrypt
python-dotenv
flasgger
opencv-python
numpy
pillow
